Move event fabric machinery into internal/api/fabric package
Only the bean post-processor stays directly under internal, alongside the
other lifecycle post-processors.
