+++ kinotic-core/src/main/java/org/kinotic/core/internal/api/fabric/BeanWiring.java
@@ -1,4 +1,4 @@
-package org.kinotic.core.internal;
+package org.kinotic.core.internal.api.fabric;
 
 import reactor.core.Disposable;
