+++ kinotic-core/src/main/java/org/kinotic/core/internal/api/fabric/MethodTarget.java
@@ -1,4 +1,4 @@
-package org.kinotic.core.internal;
+package org.kinotic.core.internal.api.fabric;
 
 import java.lang.reflect.Method;
